test: split the integration tests into smaller tests

+++ scenario_factory/pipeline.py
@@ -90,11 +90,13 @@ class PipelineContext:
 # Type aliases to make the function definitions more readable
 _PipelineMapFuncType: TypeAlias = Callable[[PipelineContext, _PipelineStepInputTypeT], _PipelineStepOutputTypeT]
 _PipelineMapFuncWithArgsType: TypeAlias = Callable[
-    [_PipelineStepArgumentsTypeT, PipelineContext, _PipelineStepInputTypeT], _PipelineStepOutputTypeT
+    [_PipelineStepArgumentsTypeT, PipelineContext, _PipelineStepInputTypeT],
+    _PipelineStepOutputTypeT,
 ]
 
 _PipelineFoldFuncType: TypeAlias = Callable[
-    [PipelineContext, Sequence[_PipelineStepInputTypeT]], Sequence[_PipelineStepOutputTypeT]
+    [PipelineContext, Sequence[_PipelineStepInputTypeT]],
+    Sequence[_PipelineStepOutputTypeT],
 ]
 
 
@@ -119,16 +121,29 @@ class PipelineStepType(Enum):
 
 class PipelineStepMode(Enum):
     CONCURRENT = auto()
+    """Run the step in a semi-parellel manner, by distributing the indidivual tasks to different threads"""
+
     PARALLEL = auto()
+    """Run this step in a true parallel manner, by distributing the individual tasks to different processes"""
+
     SEQUENTIAL = auto()
+    """Run this step sequentially on the main thread"""
 
 
-_AnyPipelineStep = TypeVar("_AnyPipelineStep", _PipelineMapFuncType, _PipelineFilterFuncType, _PipelineFoldFuncType)
+_AnyPipelineStep = TypeVar(
+    "_AnyPipelineStep",
+    _PipelineMapFuncType,
+    _PipelineFilterFuncType,
+    _PipelineFoldFuncType,
+)
 
 
 class PipelineStep(Generic[_AnyPipelineStep]):
     def __init__(
-        self, step_func: _AnyPipelineStep, type: PipelineStepType, mode: PipelineStepMode = PipelineStepMode.CONCURRENT
+        self,
+        step_func: _AnyPipelineStep,
+        type: PipelineStepType,
+        mode: PipelineStepMode = PipelineStepMode.CONCURRENT,
     ):
         self._step_func: _AnyPipelineStep = step_func
         self._type = type
@@ -188,7 +203,7 @@ def pipeline_map(
 
 def pipeline_map_with_args(
     mode: PipelineStepMode = PipelineStepMode.CONCURRENT,
-) -> Callable[[_PipelineMapFuncWithArgsType], Callable[[PipelineStepArguments], PipelineStep[_PipelineMapFuncType]]]:
+) -> Callable[[_PipelineMapFuncWithArgsType], Callable[[PipelineStepArguments], PipelineStep[_PipelineMapFuncType]],]:
     """
     Decorate a function to indicate its use as a map function for the pipeline. This decorator will partially apply the function by setting the args parameter.
     """
@@ -200,7 +215,11 @@ def pipeline_map_with_args(
             args: PipelineStepArguments,
         ) -> PipelineStep[_PipelineMapFuncType]:
             step_func_with_args_applied = functools.partial(func, args)
-            return PipelineStep(step_func=step_func_with_args_applied, type=PipelineStepType.MAP, mode=mode)
+            return PipelineStep(
+                step_func=step_func_with_args_applied,
+                type=PipelineStepType.MAP,
+                mode=mode,
+            )
 
         return inner_wrapper
 
@@ -231,7 +250,11 @@ def pipeline_filter(
             filter: PipelineFilterPredicate,
         ) -> PipelineStep[_PipelineFilterFuncType]:
             step_func_with_args_applied = functools.partial(func, filter)
-            return PipelineStep(step_func=step_func_with_args_applied, type=PipelineStepType.FILTER, mode=mode)
+            return PipelineStep(
+                step_func=step_func_with_args_applied,
+                type=PipelineStepType.FILTER,
+                mode=mode,
+            )
 
         return inner_wrapper
 
@@ -275,7 +298,10 @@ def _wrap_pipeline_step(
 
 
 def _execute_pipeline_step(
-    ctx: PipelineContext, pipeline_step, step_index: int, input_value: _PipelineStepInputTypeT
+    ctx: PipelineContext,
+    pipeline_step,
+    step_index: int,
+    input_value: _PipelineStepInputTypeT,
 ) -> Tuple[int, PipelineStepResult[_PipelineStepInputTypeT, _PipelineStepOutputTypeT]]:
     result: PipelineStepResult[_PipelineStepInputTypeT, _PipelineStepOutputTypeT] = _wrap_pipeline_step(
         ctx, pipeline_step, input_value
@@ -292,7 +318,21 @@ def _process_worker_init(seed: int) -> None:
 
 
 class PipelineExecutor:
-    def __init__(self, ctx: PipelineContext, steps: List[PipelineStep], num_threads: int, num_processes: int):
+    """Execute the steps in a pipeline
+
+    :param ctx:
+    :param steps: The list of pipeline steps that should be executed. Order dependent.
+    :param num_threads: If provided, enables the concurrent execution of pipeline steps on a thread pool with :param:`num_threads` worker threads
+    :param num_processes: If provided, enables the parallel execution of pipeline steps on a process pool with :param:`num_processes` worker processes
+    """
+
+    def __init__(
+        self,
+        ctx: PipelineContext,
+        steps: List[PipelineStep],
+        num_threads: Optional[int] = None,
+        num_processes: Optional[int] = None,
+    ) -> None:
         self._ctx = ctx
         self._steps = steps
 
@@ -301,11 +341,25 @@ class PipelineExecutor:
         self._pipeline_step_results: List[PipelineStepResult] = []
 
         seed = ctx.get_scenario_factory_config().seed
+        random.seed(seed)
+        np.random.seed(seed)
 
-        self._process_pool = multiprocess.Pool(
-            processes=num_processes, initializer=_process_worker_init, initargs=(seed,)
-        )
-        self._thread_executor = ThreadPoolExecutor(max_workers=num_threads)
+        if num_processes is None:
+            self._parallism_enabled = False
+        else:
+            self._process_pool = multiprocess.Pool(
+                processes=num_processes,
+                initializer=_process_worker_init,
+                initargs=(seed,),
+            )
+            self._parallism_enabled = True
+
+        if num_threads is None:
+            self._concurrency_enabled = False
+        else:
+            self._thread_executor = ThreadPoolExecutor(max_workers=num_threads)
+            self._concurrency_enabled = True
+
         # By default, no tasks may be scheduled on the worker pools
         self._scheduling_enabled = False
 
@@ -318,29 +372,45 @@ class PipelineExecutor:
         if not self._scheduling_enabled:
             return
         self._num_of_running_pipeline_steps += 1
+
+        # To execute a fold step, the executor has to collect all values until the fold step first.
+        # This is done by suspending the execution of each specific value,
+        # and notifying the executor. The fold operation will then be performed
+        # in the main loop on the main thread once all values have been yield.
         if step.type == PipelineStepType.FOLD:
             return self._yield_for_fold(step, step_index, input_value)
 
-        if step.mode == PipelineStepMode.CONCURRENT:
+        if step.mode == PipelineStepMode.CONCURRENT and self._concurrency_enabled:
+            # Steps that can be executed concurrently, are submitted to the thread pool
             new_f: Future[Tuple[int, PipelineStepResult]] = self._thread_executor.submit(
                 _execute_pipeline_step, self._ctx, step, step_index, input_value
             )
             new_f.add_done_callback(self._chain_next_step_from_previous_step_future)
-        elif step.mode == PipelineStepMode.PARALLEL:
+        elif step.mode == PipelineStepMode.PARALLEL and self._parallism_enabled:
+            # Steps that must be executed in parallel, are submitted to the process pool to ensure true parallelism
             self._process_pool.apply_async(
                 _execute_pipeline_step,
                 (self._ctx, step, step_index, input_value),
                 callback=self._chain_next_step_from_previous_step_callback,
             )
         else:
-            raise NotImplementedError()
+            # Either the step's mode is `PipelineStepMode.SEQUENTIAL` or it is one of
+            # the other modes, but the mode is disabled for the executor.
+            result: Tuple[int, PipelineStepResult] = _execute_pipeline_step(self._ctx, step, step_index, input_value)
+            self._chain_next_step_from_previous_step_callback(result)
 
-    def _chain_next_step_from_previous_step_callback(self, result: Tuple[int, PipelineStepResult]):
+    def _chain_next_step_from_previous_step_callback(self, result: Tuple[int, PipelineStepResult]) -> None:
+        """
+        Create a future from the result of a pipeline step execution and chain the execution of the consecutive pipeline step.
+        """
         new_future: Future[Tuple[int, PipelineStepResult]] = Future()
         new_future.set_result(result)
         self._chain_next_step_from_previous_step_future(new_future)
 
     def _chain_next_step_from_previous_step_future(self, future: Future[Tuple[int, PipelineStepResult]]) -> None:
+        """
+        Handles the result of a pipeline step execution and executes the consecutive pipeline step, if applicable.
+        """
         # If this method is called, this means that a previous task has finished executing
         self._num_of_running_pipeline_steps -= 1
 
@@ -416,7 +486,12 @@ class PipelineExecutor:
         self._num_of_running_pipeline_steps = 1
         # The fold will be simply executed sequentially on the main thread in the main loop.
         # Although, it could be submitted to the worker pool, there does not seem to be any benefit from doing so
-        result = _execute_pipeline_step(self._ctx, self._fold_step, self._fold_step_index, self._values_queued_for_fold)
+        result = _execute_pipeline_step(
+            self._ctx,
+            self._fold_step,
+            self._fold_step_index,
+            self._values_queued_for_fold,
+        )
 
         # Reset the fold state, *before* the next tasks are scheduled.
         # This is done, so that no race-condition is encountered if the
@@ -452,9 +527,11 @@ class PipelineExecutor:
         finally:
             # make sure that no new tasks will be scheduled during shutdown
             self._scheduling_enabled = False
-            self._thread_executor.shutdown()
-            self._process_pool.terminate()
-            self._process_pool.close()
+            if self._concurrency_enabled:
+                self._thread_executor.shutdown()
+            if self._parallism_enabled:
+                self._process_pool.terminate()
+                self._process_pool.close()
 
         return self._pipeline_step_results
 
@@ -483,7 +560,9 @@ class PipelineExecutionResult:
         for pipeline_step, cum_time_ns in cum_time_by_pipeline_step.items():
             print(
                 fmt_str.format(
-                    pipeline_step, round(cum_time_ns / 1000000000, 2), cum_elements_by_pipeline_step[pipeline_step]
+                    pipeline_step,
+                    round(cum_time_ns / 1000000000, 2),
+                    cum_elements_by_pipeline_step[pipeline_step],
                 )
             )
 
@@ -569,16 +648,16 @@ class Pipeline:
         self,
         input_values: Iterable,
         ctx: PipelineContext,
-        num_threads: Optional[int] = None,
-        num_processes: Optional[int] = None,
+        num_threads: Optional[int] = multiprocess.cpu_count(),
+        num_processes: Optional[int] = multiprocess.cpu_count(),
     ) -> PipelineExecutionResult:
         """
         Execute the pipeline on the :param:`input_values` with :param:`ctx`.
 
         :param input_values: An iterable containing the input values for the first step in the pipeline
         :param ctx: The pipeline context for this specific execution
-        :param num_threads: Configure the number of threads in the worker pool
-        :param num_processes: Configure the number of processes in the worker pool
+        :param num_threads: Configure the number of threads in the worker pool. If None is provided, multithreading will be disabled.
+        :param num_processes: Configure the number of processes in the worker pool. If None is provided, multiprocessing will be disabled.
 
         :returns: The result of the execution
         """
@@ -591,16 +670,11 @@ class Pipeline:
             raise RuntimeError(
                 "Cannot execute pipeline: pipeline has duplicated steps! Each pipeline step might only occur once in a single pipeline!"
             )
-        if num_threads is None:
-            num_threads = multiprocess.cpu_count()
 
-        if num_processes is None:
-            num_processes = multiprocess.cpu_count()
-
-        if num_threads < 1:
+        if num_threads is not None and num_threads < 1:
             raise ValueError("Number of threads for pipeline execution must be at least 1")
 
-        if num_processes < 1:
+        if num_processes is not None and num_processes < 1:
             raise ValueError("Number of processes for pipeline execution must be at least 1")
 
         start_time = time.time_ns()
